Refactor: Moved all Dto's to Model/Request or Response, renamed a some.
