--- Server/JokesOnYou.Web.Api/DTOs/FlagReplyDto.cs
@@ -1,6 +1,0 @@
-﻿namespace JokesOnYou.Web.Api.DTOs
-{
-    public class FlagReplyDto
-    {
-    }
-}
