--- Server/JokesOnYou.Web.Api/DTOs/TagUserDTO.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace JokesOnYou.Web.Api.DTOs
-{
-    public class TagUserDTO
-    {
-        public int Id { get; set; }
-        public string Role { get; set; }
-    }
-}
