+++ Server/JokesOnYou.Web.Api/JokesOnYou.Web.Api.xml
@@ -4,7 +4,7 @@
         <name>JokesOnYou.Web.Api</name>
     </assembly>
     <members>
-        <member name="T:JokesOnYou.Web.Api.DTOs.JokeUpdateDto">
+        <member name="T:JokesOnYou.Web.Api.Models.Request.JokeUpdateDto">
             <summary>
             Does not update joke's id, likes, author, uploaddate, or dislikes. Just Premise and Punchline
             </summary>
