+++ Server/JokesOnYou.Web.Api/Models/Request/ReasonCreateDto.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace JokesOnYou.Web.Api.DTOs
+namespace JokesOnYou.Web.Api.Models.Request
 {
     public class ReasonCreateDto
     {
