+++ Server/JokesOnYou.Web.Api/Models/Response/FlagReplyDto.cs
@@ -1,0 +1,6 @@
+﻿namespace JokesOnYou.Web.Api.Models.Response
+{
+    public class FlagReplyDto
+    {
+    }
+}
