+++ Server/JokesOnYou.Web.Api/Models/Response/LikedJokeReplyDto.cs
@@ -1,4 +1,4 @@
-﻿namespace JokesOnYou.Web.Api.DTOs
+﻿namespace JokesOnYou.Web.Api.Models.Response
 {
     public class LikedJokeReplyDto
     {
