+++ Server/JokesOnYou.Web.Api/Models/Response/TagUserDTO.cs
@@ -1,0 +1,8 @@
+﻿namespace JokesOnYou.Web.Api.Models.Response
+{
+    public class TagUserDTO
+    {
+        public int Id { get; set; }
+        public string Role { get; set; }
+    }
+}
