+++ Server/JokesOnYou.Web.Api/Profiles/JokesProfile.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
-using JokesOnYou.Web.Api.DTOs;
+using JokesOnYou.Web.Api.Models.Request;
 using JokesOnYou.Web.Api.Models;
+using JokesOnYou.Web.Api.Models.Response;
 
 namespace JokesOnYou.Web.Api.Profiles
 {
@@ -9,7 +10,6 @@ namespace JokesOnYou.Web.Api.Profiles
         public JokesProfile()
         {
             CreateMap<Joke, JokeReplyDto>();
-            CreateMap<User, JokeAuthorDto>();
             CreateMap<JokeCreateDto, Joke>();
             CreateMap<JokeUpdateDto, Joke>();
         }
