+++ Server/JokesOnYou.Web.Api/Profiles/TagProfile.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
-using JokesOnYou.Web.Api.DTOs;
+using JokesOnYou.Web.Api.Models.Request;
 using JokesOnYou.Web.Api.Models;
+using JokesOnYou.Web.Api.Models.Response;
 
 namespace JokesOnYou.Web.Api.Profiles
 {
