+++ Server/JokesOnYou.Web.Api/Profiles/UsersProfile.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
-using JokesOnYou.Web.Api.DTOs;
+using JokesOnYou.Web.Api.Models.Request;
 using JokesOnYou.Web.Api.Models;
+using JokesOnYou.Web.Api.Models.Response;
 
 namespace JokesOnYou.Web.Api.Profiles
 {
@@ -12,6 +13,7 @@ namespace JokesOnYou.Web.Api.Profiles
             CreateMap<UserRegisterDTO, User>();
             CreateMap<UserUpdateDTO, User>();
             CreateMap<User, UserReplyDTO>();
+            CreateMap<User, JokeAuthorDto>();
         }
     }
 }
