+++ Server/JokesOnYou.Web.Api/Repositories/Interfaces/IUserRepository.cs
@@ -1,7 +1,8 @@
-﻿using JokesOnYou.Web.Api.DTOs;
+﻿using JokesOnYou.Web.Api.Models.Request;
 using JokesOnYou.Web.Api.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using JokesOnYou.Web.Api.Models.Response;
 
 namespace JokesOnYou.Web.Api.Repositories.Interfaces
 {
